Move errors to separate file
